reactor: Move update_reactor to its own file to prepare for the realtime system and reduce the amount of code in main that students don't need to look at.
