+++ common.c
@@ -1,5 +1,6 @@
 #include "common.h"
 
+exit_reason_t exit_reason = exit_reason_none;
 char rod_depth = 16;
 float coolant_flow = 10;
 float coolant_temp = 70.0;
