+++ status_win.h
@@ -3,7 +3,7 @@
 #define STATUS_WIN_X 0
 #define STATUS_WIN_Y 0
 #define STATUS_WIN_W 80
-#define STATUS_WIN_H 10
+#define STATUS_WIN_H 9
 
 void status_init(void);
