+++ wormwood.c
@@ -8,40 +8,12 @@
 #include "common.h"
 #include "console_win.h"
 #include "status_win.h"
-
-#define DEBUG 0
-#define FUNCNAME if (DEBUG) console_printf("In %s @ line %d...\n", __func__, __LINE__);
-
-#define MAX_SAFE_DEPTH 16
-#define MAX_FLOW_RATE 100.0
-
-typedef enum {
-	exit_reason_none = 0,
-	exit_reason_fail,
-	exit_reason_quit,
-} exit_reason_t;
-
-exit_reason_t exit_reason = exit_reason_none;
+#include "reactor.h"
 
 void get_string(char *dest) {
 	char input_buffer[8192];
 	console_read_str(input_buffer); // Creates a new vuln, do we want this?
 	strcpy(dest, input_buffer);
-}
-
-void print_sparks(void) {
-	int i = CONSOLE_WIN_W * 3;
-	int j = 0;
-	while (i > 0) {
-		for (j = (rand() % 10); j > 0 && i > 1; j--) {
-			console_printf(" ");
-			i--;
-		}
-		console_printf("*");
-		i--;
-	}
-
-	console_printf("\n");
 }
 
 void auth_user(void) {
@@ -202,114 +174,6 @@ void get_and_do_choice(void) {
 	return;
 }
 
-float float_up_to(int max) {
-	// https://stackoverflow.com/questions/13408990/how-to-generate-random-float-number-in-c
-	float fuzz = (float)rand()/(float)(RAND_MAX/max);
-	return fuzz;
-}
-
-float get_fuzz(void) {
-	return float_up_to(1);
-}
-
-int rand_sign(void) {
-	int sign;
-	int randval = rand();
-	if (randval % 2 == 0) {
-		sign = 1;
-	}
-	else {
-		sign = -1;
-	}
-
-	return sign;
-}
-
-void update_reactor(void) {
-	FUNCNAME
-
-	/* DO NOT UPDATE THIS CODE */
-
-	/* This may shock you, but this is not a real nuclear reactor.
-	 * the physics of this "simulation" are not even remotely accurate...
-	 * in fact, I'm just making them up.
-	 * But, for the purposes of this activity, pretend that there is
-	 * real danger or stakes. It's more engaging and interesting that way!
-	 */
-
-	/* WARNINGS */
-	if (reactor_temp > 3000) {
-		console_printf("\n***** WARNING: REACTOR COOLANT WILL VAPORIZE AT 5000 DEGREES ******\n");
-		if(reactor_temp > 4000) {
-			console_printf("***** WARNING: IMMINENT BREACH! IMMINENT BREACH! ******\n");
-		}
-		console_printf("\n");
-	}
-
-	/* Fail/throw error if reactor temp goes over 5K. */
-	if (reactor_temp >= 5000) {
-		console_clear();
-		print_sparks();
-		console_printf("****** COOLANT VAPORIZATION *******\n");
-		console_printf("****** CONTAINMENT VESSEL VENTING *******\n");
-		console_printf("****** MAJOR RADIOACTIVITY LEAK!!! *******\n\n");
-		print_sparks();
-		exit_reason = exit_reason_fail;
-		return;
-	}
-
-	/* COOLANT FLOW HEAT REDUCTION. */
-	/* Coolant flow reduces reactor temp. */
-	if (reactor_temp > 70) {
-		reactor_temp = reactor_temp - ((coolant_flow * float_up_to(7)));
-	}
-
-	/* Coolant temp follows the reactor temp, but at a delay. */
-	coolant_temp = coolant_temp + ((reactor_temp - coolant_temp) * .15);
-
-	/* Reactor cannot get below room temp. */
-	if (reactor_temp < 70) {
-		reactor_temp = 70;
-	}
-
-	/* Fuzz the reactor temp a bit for realism. */
-	reactor_temp = reactor_temp + get_fuzz() * rand_sign();
-
-	/* RETRACTING THE RODS INCREASES THE TEMP */
-	/* for each unit the rod is not fully extracted, add a random float up to 50. */
-	float bump = 0;
-	int i;
-	int rod_factor = MAX_SAFE_DEPTH - rod_depth;
-
-	for (i = rod_factor; i > 0; i--) {
-		bump = bump + float_up_to(20);
-	}
-	reactor_temp = reactor_temp + bump;
-
-	/* SAFETY PROTOCOLS */
-	if (safety_enabled == true && reactor_temp > 2000) {
-		safety_active = 1;
-		console_printf("\n ****** SAFETY PROTOCOLS ENGAGED: Extending control rods! *******\n\n");
-		if (rod_depth <= MAX_SAFE_DEPTH) {
-			/* Automatically increment rod_depth to cool reactor. */
-			rod_depth++;
-		}
-
-		console_printf("\n ****** SAFETY PROTOCOLS ENGAGED: Increasing coolant flow! *******\n\n");
-		if (coolant_flow <= MAX_FLOW_RATE) {
-			coolant_flow = coolant_flow + 1;
-			if (coolant_flow > MAX_FLOW_RATE) {
-				coolant_flow = MAX_FLOW_RATE;
-			}
-		}
-	}
-
-	if (safety_active == 1 && reactor_temp < 2000) {
-		safety_active = 0;
-		console_printf("\n\n******* NORMAL OPERATING TEMPERATURE ACHIEVED ********\n\n");
-	}
-}
-
 void reactor_status(void) {
 	FUNCNAME
 
@@ -323,23 +187,10 @@ void reactor_status(void) {
 	get_and_do_choice();
 	update_reactor();
 
-	/* Check if rod depth is safe. */
-	if (rod_depth < 0 || rod_depth > MAX_SAFE_DEPTH) {
-		console_clear();
-		print_sparks();
-		console_printf("WARNING! WARNING! WARNING!\n");
-		console_printf("CONTAINMENT VESSEL RUPTURE!\n");
-		console_printf("CONTROL RODS EXTENDED THROUGH CONTAINMENT VESSEL!!!\n");
-	    console_printf("RADIATION LEAK - EVACUATE THE AREA!\n\n");
-		print_sparks();
-		exit_reason = exit_reason_fail;
-	}
-
 	console_wait_until_press();
 
 	return;
 }
-
 
 int main(int argc, char *argv[]) {
 	/* Initialize ncurses. */
